Module rest::basis_io

Modules

basic_math 🔒

basis_list

bse_downloader

etb

Structs

BasCell

#BasCell BasCell has the same organization structure of the basis set files from the basis set exchange (BSE) BasCell.function_type: “gto_spherical” or “gto” BasCell.region:
BasCell.angular_momentum: The angular momentums of the GTOs in this cell BasCell.exponents: The exponents BasCell.coefficients: The coefficients

BasCellRaw

BasInfo

BasInfo

Basis4Elem

#Basis4Elem Basis4Elem contains all informations of the GTO basis functions located in a given element self.electron_shells: the GTO basis functions organized cell by cell BasCell self.references: The reference of each basis cell. self.global_index: The global index for the given atom

Basis4ElemRaw

RefCell

Functions

_gaussian_int

cartesian_gto_1st_batch

cartesian_gto_1st_batch_par

cartesian_gto_1st_batch_par_v02

cartesian_gto_1st_batch_v02

cartesian_gto_1st_batch_v03

cartesian_gto_1st_cint

cartesian_gto_1st_cint_batch

cartesian_gto_1st_cint_batch_serial

cartesian_gto_1st_cint_batch_serial_v03

cartesian_gto_1st_std

cartesian_gto_batch

cartesian_gto_batch_par

cartesian_gto_batch_par_v02

cartesian_gto_batch_par_v03

cartesian_gto_batch_v03

cartesian_gto_batch_v04

cartesian_gto_cint

cartesian_gto_cint_batch

cartesian_gto_cint_batch_rayon

cartesian_gto_cint_batch_v04

cartesian_gto_std

cint_norm_factor

Normalization factor for GTO radial part g=r^l e^{-\alpha r^2}

gto_1st_value

gto_value

gto_value_debug

spheric_gto_1st_value_batch

spheric_gto_1st_value_batch_serial

spheric_gto_1st_value_serial

spheric_gto_value_matrixfull

spheric_gto_value_matrixfull_serial

spheric_gto_value_serial